+++ TaxCalculator/ITaxCalculation.cs
@@ -2,6 +2,6 @@ namespace TaxCalculator
 {
     public interface ITaxCalculation
     {
-        double Calculate(double amount, ITaxSettings settings);
+        decimal Calculate(decimal amount, ITaxSettings settings);
     }
 }
