+++ TaxCalculator/ProgressiveTaxCalculation.cs
@@ -4,21 +4,21 @@ namespace TaxCalculator
 {
     public class ProgressiveTaxCalculation : ITaxCalculation
     {
-        public double Calculate(double amount, ITaxSettings settings)
+        public decimal Calculate(decimal amount, ITaxSettings settings)
         {
             //we expect a table of values in the format: from:to
             //the from starts from 0 and extends to the max value (100000000)
 
-            double currentMin = 0;
-            double currentTax = 0;
+            decimal currentMin = 0;
+            decimal currentTax = 0;
             foreach (var bracket in settings.GetList())
             {
                 if (amount < currentMin)
                     break;
                 try
                 {
-                    double upperBound = Convert.ToDouble(bracket.name);
-                    double taxPerc = Convert.ToDouble(bracket.value);
+                    decimal upperBound = Convert.ToDecimal(bracket.name);
+                    decimal taxPerc = Convert.ToDecimal(bracket.value);
                     var bracketAmount = Math.Min(amount, upperBound) - currentMin;
                     var taxAmount = bracketAmount * taxPerc;
                     currentTax += taxAmount;
